ft_substr && libft updated

+++ libft.a/ft_strlen.c
@@ -1,7 +1,7 @@
 #include	"libft.h"
 #include	<stdio.h>
 
-int	ft_strlen(const char	*str)
+size_t ft_strlen(const char	*str)
 {
 	int len;
